# Speeding up b.clear()