This PR speeds up `b.clear()` by avoiding the internal use of `b.forEach()`. A benchmark script `clearItems.jsx` is added to `test/benchmarks/`.

On my system the benchmark gave these results:

```
Old b.clear():
Clearing 500 grouped items: 8.066 seconds.
Clearing 500 items on page: 4.259 seconds.
Clearing 500 items on layer: 6.324 seconds.
Clearing 500 items in document: 2.749 seconds.
```

```
New b.clear():
Clearing 500 grouped items: 0.275 seconds.
Clearing 500 items on page: 0.589 seconds.
Clearing 500 items on layer: 0.561 seconds.
Clearing 500 items in document: 0.453 seconds.
```

The more objects need to be deleted, the greater the speed difference. It is also greater if the objects are deleted from the active page as the page needs to be redrawn more often with the old `b.clear()`.

It passes all tests.